+++ Lawn/System/PlayerInfo.h
@@ -49,7 +49,7 @@ public:
     int                 mCoins;                             //+0x28
     int                 mFinishedAdventure;                 //+0x2C
     int                 mChallengeRecords[100];             //+0x30
-    int                 mPurchases[80];                     //+0x1C0
+    long                mPurchases[80];                     //+0x1C0
     int                 mPlayTimeActivePlayer;              //+0x300
     int                 mPlayTimeInactivePlayer;            //+0x304
     int                 mHasUsedCheatKeys;                  //+0x308
